+++ blockfrost-api/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Allow servant `0.20`
 * Additions
   * `/accounts/:stake_address/addresses/total` endpoint and `AddressAssociatedTotal` type
+  * `/pools/extended` endpoint and `Pool` type
 
 # Version [0.8.1.0](https://github.com/blockfrost/blockfrost-haskell/compare/api-0.8.0.0...api-0.8.1.0) (2023-09-18)